MNT:Move common routines to amutils.py
