+++ amostra/client/commands.py
@@ -2,83 +2,16 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 from doct import Document
 import ujson
-from uuid import uuid4
 import requests
-import time as ttime
 from ..client import conf
-from .amutils import doc_or_uid_to_uid
+from .amutils import doc_or_uid_to_uid, _get, _post, _put
 
 # TODO: Add tests for both local and online commands/clients
 # TODO: Add AmostraClient as a convenient way to interact with commands.py
 
+
 class AmostraException(Exception):
     pass
-
-
-def _get(url, params):
-    """RESTful API get (querying)
-
-    Parameters
-    ----------
-    url: str
-        Address for the amostra server
-    params: dict
-        Query parameters to be sent to mongo instance
-
-    Returns
-    -------
-    list
-        Results of the query
-
-    """
-    r = requests.get(url, ujson.dumps(params))
-    r.raise_for_status()
-    return ujson.loads(r.text)
-
-
-def _post(url, data):
-    """RESTful API post (insert to database)
-
-    Parameters
-    ----------
-    url: str
-        Address for the amostra server
-    data: dict
-        Entries to be inserted to database
-
-    """
-    r = requests.post(url,
-                      data=ujson.dumps(data))
-    r.raise_for_status()
-    return r.json()
-
-
-def _put(url, query, update):
-    """RESTful API put (update entries in database)
-
-    Parameters
-    ----------
-    url: str
-        Address for the amostra server
-    query: dict
-        Query string. Any pymongo supported mongo query
-    update: dict
-        Key/value pairs to be updated in the original document
-
-    Returns
-    -------
-    bool
-        True if update successful
-    Raises
-    --------
-    requests.exceptions.HTTPError
-        In case update fails, appropriate HTTPError and message string returned
-
-    """
-    update_cont = {'query': query, 'update': update}
-    r = requests.put(url,
-                     data=ujson.dumps(update_cont))
-    r.raise_for_status()
 
 
 class SampleReference(object):
